Split ApiClient from Github
